fix lint

+++ src/graph/network.h
@@ -15,9 +15,9 @@ namespace network {
 // TODO(chao): make these numbers configurable
 
 // Each single message cannot larger than 300 MB
-const long long kMaxBufferSize = 300 * 1024 * 2014;
+const int64 kMaxBufferSize = 300 * 1024 * 2014;
 // Size of message queue is 1 GB
-const long long kQueueSize = 1024 * 1024 * 1024;
+const int64 kQueueSize = 1024 * 1024 * 1024;
 // Maximal try count of connection
 const int kMaxTryCount = 500;
